fix: align feature paths and reorganize patents styles
