+++ src/features/Home/Routes/index.ts
@@ -1,3 +1,3 @@
 export { default as Home } from '../Home';
-export { default as AboutMe } from '../Components/AboutMe';
-export { default as Experience } from '../Components/Experience';
+export { default as AboutMe } from '../components/AboutMe';
+export { default as Experience } from '../components/Experience';
