+++ src/features/Home/components/Experience.tsx
@@ -1,7 +1,7 @@
 import '../styles/AboutMeSpeechBubble.css';
 import '../styles/ExperienceLiveIcon.css';
 import { useState } from 'react';
-import ExperienceModal from '../Components/ExperienceModal/ExperienceModal';
+import ExperienceModal from '../components/ExperienceModal/ExperienceModal';
 import { experiences } from '../helpers/experienceData';
 import type { ExperienceType } from '../helpers/experienceData';
 import { useTranslation } from 'react-i18next';
