+++ src/features/NoPage/Routes/index.ts
@@ -1,1 +1,1 @@
-export { default as NoPage } from '../Components/NoPage';
+export { default as NoPage } from '../components/NoPage';
