+++ src/features/Patents/styles/PatentsHero.css
