+++ src/features/Technologies/Routes/index.ts
@@ -1,1 +1,1 @@
-export { default as Technologies } from '../Components/Technologies';
+export { default as Technologies } from '../components/Technologies';
